requirements.txt
CTkSpinbox==1.4.1
customtkinter==5.2.2
Jinja2==3.1.4
mysql-connector-python==8.4.0
pillow==10.3.0
PyAutoGUI==0.9.54
requests== 2.32.3
tkcalendar==1.6.1

